pyproject.toml
[project]
name = "seg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "scipy>=1.15.3",
    "torch>=2.10.0",
    "transformers>=5.2.0",
]
